a little change

+++ Voting15_Online/ext.h
@@ -1,8 +1,8 @@
 #ifndef EXT_H
 #define EXT_H
 
-extern const int RINF;
-extern const int INF;
+extern const double RINF;
+extern const double INF;
 extern int APPNUM;
 extern int N;
 extern int M;
